HDF5

Base.isopen

HDF5.create_external

HDF5.get_datasets

HDF5.get_dims

HDF5.h5open

HDF5.h5open

HDF5.ishdf5

HDF5.set_dims!

HDF5.start_swmr_write

Base.isopen — Method

isopen(obj::HDF5File)

Returns true if obj has not been closed, false if it has been closed.

HDF5.create_external — Method

create_external(source::Union{HDF5File, HDF5Group}, source_relpath, target_filename, target_path;
                lcpl_id=HDF5.H5P_DEFAULT, lapl_id=HDF5.H5P.DEFAULT)

Create an external link such that source[source_relpath] points to target_path within the file with path target_filename; Calls [H5Lcreate_external](https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-CreateExternal).

HDF5.get_datasets — Method

get_datasets(file::HDF5File) -> datasets::Vector{HDF5Dataset}

Get all the datasets in an hdf5 file without loading the data.

HDF5.get_dims — Method

get_dims(dset::HDF5Dataset)

Get the array dimensions from a dataset and return a tuple of dims and maxdims.

HDF5.h5open — Function

h5open(filename::AbstractString, mode::AbstractString="r"; swmr=false)

Open or create an HDF5 file where mode is one of:

"r" read only

"r+" read and write

"cw" read and write, create file if not existing, do not truncate

"w" read and write, create a new file (destroys any existing contents)

Pass swmr=true to enable (Single Writer Multiple Reader) SWMR write access for "w" and "r+", or SWMR read access for "r".

HDF5.h5open — Method

function h5open(f::Function, args...; swmr=false)

Apply the function f to the result of h5open(args...;kwargs...) and close the resulting HDF5File upon completion. For example with a do block:

h5open("foo.h5","w") do h5
    h5["foo"]=[1,2,3]
end

HDF5.ishdf5 — Method

ishdf5(name::AbstractString)

Returns true if name is a path to a valid hdf5 file, false otherwise.

HDF5.set_dims! — Method

set_dims!(dset::HDF5Dataset, new_dims::Dims)

Change the current dimensions of a dataset to new_dims, limited by max_dims = get_dims(dset)[2]. Reduction is possible and leads to loss of truncated data.

HDF5.start_swmr_write — Method

start_swmr_write(h5::HDF5File)

Start Single Reader Multiple Writer (SWMR) writing mode. See SWMR documentation.