Low-level library bindings

At the lowest level, HDF5.jl operates by calling the public API of the HDF5 shared library through a set of ccall wrapper functions. This page documents the function names and nominal C argument types of the API which have bindings in this package. Note that in many cases, high-level data types are valid arguments through automatic ccall conversions. For instance, HDF5.Datatype objects will be automatically converted to their hid_t ID by Julia's cconvert+unsafe_convert ccall rules.

There are additional helper wrappers (often for out-argument functions) which are not documented here.

H5 — General Library Functions

h5_close

h5_dont_atexit

h5_free_memory

h5_garbage_collect

h5_get_libversion

h5_is_library_threadsafe

h5_open

h5_set_free_list_limits

HDF5.API.h5_close — Function

h5_close()

See libhdf5 documentation for H5close.

HDF5.API.h5_dont_atexit — Function

h5_dont_atexit()

See libhdf5 documentation for H5dont_atexit.

HDF5.API.h5_free_memory — Function

h5_free_memory(buf::Ptr{Cvoid})

See libhdf5 documentation for H5free_memory.

HDF5.API.h5_garbage_collect — Function

h5_garbage_collect()

See libhdf5 documentation for H5garbage_collect.

HDF5.API.h5_get_libversion — Function

h5_get_libversion(majnum::Ref{Cuint}, minnum::Ref{Cuint}, relnum::Ref{Cuint})

See libhdf5 documentation for H5get_libversion.

HDF5.API.h5_is_library_threadsafe — Function

h5_is_library_threadsafe(is_ts::Ref{Cuint})

See libhdf5 documentation for H5is_library_threadsafe.

HDF5.API.h5_open — Function

h5_open()

See libhdf5 documentation for H5open.

HDF5.API.h5_set_free_list_limits — Function

h5_set_free_list_limits(reg_global_lim::Cint, reg_list_lim::Cint, arr_global_lim::Cint, arr_list_lim::Cint, blk_global_lim::Cint, blk_list_lim::Cint)

See libhdf5 documentation for H5set_free_list_limits.

H5A — Attribute Interface

h5a_close

h5a_create

h5a_create_by_name

h5a_delete

h5a_delete_by_idx

h5a_delete_by_name

h5a_exists

h5a_exists_by_name

h5a_get_create_plist

h5a_get_name

h5a_get_name_by_idx

h5a_get_space

h5a_get_type

h5a_iterate

h5a_open

h5a_read

h5a_write

HDF5.API.h5a_close — Function

h5a_close(id::hid_t)

See libhdf5 documentation for H5Aclose.

HDF5.API.h5a_create — Function

h5a_create(loc_id::hid_t, pathname::Ptr{UInt8}, type_id::hid_t, space_id::hid_t, acpl_id::hid_t, aapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Acreate2.

HDF5.API.h5a_create_by_name — Function

h5a_create_by_name(loc_id::hid_t, obj_name::Ptr{UInt8}, attr_name::Ptr{UInt8}, type_id::hid_t, space_id::hid_t, acpl_id::hid_t, aapl_id::hid_t, lapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Acreate_by_name.

HDF5.API.h5a_delete — Function

h5a_delete(loc_id::hid_t, attr_name::Ptr{UInt8})

See libhdf5 documentation for H5Adelete.

HDF5.API.h5a_delete_by_idx — Function

h5a_delete_by_idx(loc_id::hid_t, obj_name::Ptr{UInt8}, idx_type::Cint, order::Cint, n::hsize_t, lapl_id::hid_t)

See libhdf5 documentation for H5Adelete_by_idx.

HDF5.API.h5a_delete_by_name — Function

h5a_delete_by_name(loc_id::hid_t, obj_name::Ptr{UInt8}, attr_name::Ptr{UInt8}, lapl_id::hid_t)

See libhdf5 documentation for H5Adelete_by_name.

HDF5.API.h5a_exists — Function

h5a_exists(obj_id::hid_t, attr_name::Ptr{UInt8}) -> Bool

See libhdf5 documentation for H5Aexists.

HDF5.API.h5a_exists_by_name — Function

h5a_exists_by_name(loc_id::hid_t, obj_name::Ptr{UInt8}, attr_name::Ptr{UInt8}, lapl_id::hid_t) -> Bool

See libhdf5 documentation for H5Aexists_by_name.

HDF5.API.h5a_get_create_plist — Function

h5a_get_create_plist(attr_id::hid_t) -> hid_t

See libhdf5 documentation for H5Aget_create_plist.

HDF5.API.h5a_get_name — Function

h5a_get_name(attr_id::hid_t, buf_size::Csize_t, buf::Ptr{UInt8}) -> Cssize_t

See libhdf5 documentation for H5Aget_name.

HDF5.API.h5a_get_name_by_idx — Function

h5a_get_name_by_idx(loc_id::hid_t, obj_name::Cstring, index_type::Cint, order::Cint, idx::hsize_t, name::Ptr{UInt8}, size::Csize_t, lapl_id::hid_t) -> Cssize_t

See libhdf5 documentation for H5Aget_name_by_idx.

HDF5.API.h5a_get_space — Function

h5a_get_space(attr_id::hid_t) -> hid_t

See libhdf5 documentation for H5Aget_space.

HDF5.API.h5a_get_type — Function

h5a_get_type(attr_id::hid_t) -> hid_t

See libhdf5 documentation for H5Aget_type.

HDF5.API.h5a_iterate — Function

h5a_iterate(obj_id::hid_t, idx_type::Cint, order::Cint, n::Ptr{hsize_t}, op::Ptr{Cvoid}, op_data::Any)

See libhdf5 documentation for H5Aiterate2.

h5a_iterate(f, loc_id, idx_type, order, idx = 0) -> hsize_t

Executes h5a_iterate with the user-provided callback function f, returning the index where iteration ends.

The callback function must correspond to the signature

    f(loc::HDF5.API.hid_t, name::Ptr{Cchar}, info::Ptr{HDF5.API.H5A_info_t}) -> HDF5.API.herr_t

where a negative return value halts iteration abnormally, a positive value halts iteration successfully, and zero continues iteration.

Examples

julia> HDF5.API.h5a_iterate(obj, HDF5.API.H5_INDEX_NAME, HDF5.API.H5_ITER_INC) do loc, name, info
           println(unsafe_string(name))
           return HDF5.API.herr_t(0)
       end

HDF5.API.h5a_open — Function

h5a_open(obj_id::hid_t, pathname::Ptr{UInt8}, aapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Aopen.

HDF5.API.h5a_read — Function

h5a_read(attr_id::hid_t, mem_type_id::hid_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Aread.

HDF5.API.h5a_write — Function

h5a_write(attr_hid::hid_t, mem_type_id::hid_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Awrite.

H5D — Dataset Interface

h5d_close

h5d_create

h5d_create_anon

h5d_extend

h5d_fill

h5d_flush

h5d_gather

h5d_get_access_plist

h5d_get_chunk_info

h5d_get_chunk_info_by_coord

h5d_get_chunk_storage_size

h5d_get_create_plist

h5d_get_num_chunks

h5d_get_offset

h5d_get_space

h5d_get_space_status

h5d_get_storage_size

h5d_get_type

h5d_iterate

h5d_open

h5d_read

h5d_read_chunk

h5d_refresh

h5d_scatter

h5d_set_extent

h5d_vlen_get_buf_size

h5d_vlen_reclaim

h5d_write

h5d_write_chunk

HDF5.API.h5d_close — Function

h5d_close(dataset_id::hid_t)

See libhdf5 documentation for H5Dclose.

HDF5.API.h5d_create — Function

h5d_create(loc_id::hid_t, pathname::Ptr{UInt8}, dtype_id::hid_t, space_id::hid_t, lcpl_id::hid_t, dcpl_id::hid_t, dapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dcreate2.

HDF5.API.h5d_create_anon — Function

h5d_create_anon(loc_id::hid_t, type_id::hid_t, space_id::hid_t, dcpl_id::hid_t, dapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dcreate_anon.

HDF5.API.h5d_extend — Function

h5d_extend(dataset_id::hid_t, size::Ptr{hsize_t})

See libhdf5 documentation for H5Dextend.

HDF5.API.h5d_fill — Function

h5d_fill(fill::Ptr{Cvoid}, fill_type_id::hid_t, buf::Ptr{Cvoid}, buf_type_id::hid_t, space_id::hid_t)

See libhdf5 documentation for H5Dfill.

HDF5.API.h5d_flush — Function

h5d_flush(dataset_id::hid_t)

See libhdf5 documentation for H5Dflush.

HDF5.API.h5d_gather — Function

h5d_gather(src_space_id::hid_t, src_buf::Ptr{Cvoid}, type_id::hid_t, dst_buf_size::Csize_t, dst_buf::Ptr{Cvoid}, op::Ptr{Cvoid}, op_data::Any)

See libhdf5 documentation for H5Dgather.

HDF5.API.h5d_get_access_plist — Function

h5d_get_access_plist(dataset_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dget_access_plist.

HDF5.API.h5d_get_chunk_info — Function

h5d_get_chunk_info(dataset_id::hid_t, fspace_id::hid_t, index::hsize_t, offset::Ptr{hsize_t}, filter_mask::Ptr{Cuint}, addr::Ptr{haddr_t}, size::Ptr{hsize_t})

See libhdf5 documentation for H5Dget_chunk_info.

h5d_get_chunk_info(dataset_id, fspace_id, index)
h5d_get_chunk_info(dataset_id, index; fspace_id = HDF5.API.H5S_ALL)

Helper method to retrieve chunk information.

Returns a NamedTuple{(:offset, :filter_mask, :addr, :size), Tuple{HDF5.API.hsize_t, UInt32, HDF5.API.haddr_t, HDF5.API.hsize_t}}.

HDF5.API.h5d_get_chunk_info_by_coord — Function

h5d_get_chunk_info_by_coord(dataset_id::hid_t, offset::Ptr{hsize_t}, filter_mask::Ptr{Cuint}, addr::Ptr{haddr_t}, size::Ptr{hsize_t})

See libhdf5 documentation for H5Dget_chunk_info_by_coord.

h5d_get_chunk_info_by_coord(dataset_id, offset)

Helper method to read chunk information by coordinate. Returns a NamedTuple{(:filter_mask, :addr, :size), Tuple{UInt32, HDF5.API.haddr_t, HDF5.API.hsize_t}}.

HDF5.API.h5d_get_chunk_storage_size — Function

h5d_get_chunk_storage_size(dataset_id::hid_t, offset::Ptr{hsize_t}, chunk_nbytes::Ptr{hsize_t})

See libhdf5 documentation for H5Dget_chunk_storage_size.

h5d_get_chunk_storage_size(dataset_id, offset)

Helper method to retrieve the chunk storage size in bytes. Returns an integer of type HDF5.API.hsize_t.

HDF5.API.h5d_get_create_plist — Function

h5d_get_create_plist(dataset_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dget_create_plist.

HDF5.API.h5d_get_num_chunks — Function

h5d_get_num_chunks(dataset_id::hid_t, fspace_id::hid_t, nchunks::Ptr{hsize_t})

See libhdf5 documentation for H5Dget_num_chunks.

HDF5.API.h5d_get_offset — Function

h5d_get_offset(dataset_id::hid_t) -> haddr_t

See libhdf5 documentation for H5Dget_offset.

HDF5.API.h5d_get_space — Function

h5d_get_space(dataset_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dget_space.

HDF5.API.h5d_get_space_status — Function

h5d_get_space_status(dataset_id::hid_t, status::Ref{Cint})

See libhdf5 documentation for H5Dget_space_status.

h5d_get_space_status(dataset_id)

Helper method to retrieve the status of the dataset space. Returns a HDF5.API.H5D_space_status_t (Cint) indicating the status, see HDF5.API.H5D_SPACE_STATUS_* constants.

HDF5.API.h5d_get_storage_size — Function

h5d_get_storage_size(dataset_id::hid_t) -> hsize_t

See libhdf5 documentation for H5Dget_storage_size.

HDF5.API.h5d_get_type — Function

h5d_get_type(dataset_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dget_type.

HDF5.API.h5d_iterate — Function

h5d_iterate(buf::Ptr{Cvoid}, type_id::hid_t, space_id::hid_t, operator::Ptr{Cvoid}, operator_data::Any)

See libhdf5 documentation for H5Diterate.

HDF5.API.h5d_open — Function

h5d_open(loc_id::hid_t, pathname::Ptr{UInt8}, dapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Dopen2.

HDF5.API.h5d_read — Function

h5d_read(dataset_id::hid_t, mem_type_id::hid_t, mem_space_id::hid_t, file_space_id::hid_t, xfer_plist_id::hid_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dread.

HDF5.API.h5d_read_chunk — Function

h5d_read_chunk(dset::hid_t, dxpl_id::hid_t, offset::Ptr{hsize_t}, filters::Ptr{UInt32}, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dread_chunk.

HDF5.API.h5d_refresh — Function

h5d_refresh(dataset_id::hid_t)

See libhdf5 documentation for H5Drefresh.

HDF5.API.h5d_scatter — Function

h5d_scatter(op::Ptr{Cvoid}, op_data::Any, type_id::hid_t, dst_space_id::hid_t, dst_buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dscatter.

HDF5.API.h5d_set_extent — Function

h5d_set_extent(dataset_id::hid_t, new_dims::Ptr{hsize_t})

See libhdf5 documentation for H5Dset_extent.

HDF5.API.h5d_vlen_get_buf_size — Function

h5d_vlen_get_buf_size(dset_id::hid_t, type_id::hid_t, space_id::hid_t, buf::Ptr{hsize_t})

See libhdf5 documentation for H5Dvlen_get_buf_size.

h5d_vlen_get_buf_size(dataset_id, type_id, space_id)

Helper method to determines the number of bytes required to store the variable length data from the dataset. Returns a value of type HDF5.API.hsize_t.

HDF5.API.h5d_vlen_reclaim — Function

h5d_vlen_reclaim(type_id::hid_t, space_id::hid_t, plist_id::hid_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dvlen_reclaim.

HDF5.API.h5d_write — Function

h5d_write(dataset_id::hid_t, mem_type_id::hid_t, mem_space_id::hid_t, file_space_id::hid_t, xfer_plist_id::hid_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dwrite.

HDF5.API.h5d_write_chunk — Function

h5d_write_chunk(dset_id::hid_t, dxpl_id::hid_t, filter_mask::UInt32, offset::Ptr{hsize_t}, bufsize::Csize_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5Dwrite_chunk.

H5E — Error Interface

h5e_close_stack

h5e_get_auto

h5e_get_current_stack

h5e_get_msg

h5e_get_num

h5e_set_auto

h5e_walk

HDF5.API.h5e_close_stack — Function

h5e_close_stack(stack_id::hid_t)

See libhdf5 documentation for H5Eclose_stack.

HDF5.API.h5e_get_auto — Function

h5e_get_auto(estack_id::hid_t, func::Ref{Ptr{Cvoid}}, client_data::Ref{Ptr{Cvoid}})

See libhdf5 documentation for H5Eget_auto2.

HDF5.API.h5e_get_current_stack — Function

h5e_get_current_stack() -> hid_t

See libhdf5 documentation for H5Eget_current_stack.

HDF5.API.h5e_get_msg — Function

h5e_get_msg(mesg_id::hid_t, mesg_type::Ref{Cint}, mesg::Ref{UInt8}, len::Csize_t) -> Cssize_t

See libhdf5 documentation for H5Eget_msg.

mesg_type, mesg = h5e_get_msg(meshg_id)

HDF5.API.h5e_get_num — Function

h5e_get_num(estack_id::hid_t) -> Cssize_t

See libhdf5 documentation for H5Eget_num.

HDF5.API.h5e_set_auto — Function

h5e_set_auto(estack_id::hid_t, func::Ptr{Cvoid}, client_data::Ptr{Cvoid})

See libhdf5 documentation for H5Eset_auto2.

HDF5.API.h5e_walk — Function

h5e_walk(stack_id::hid_t, direction::Cint, op::Ptr{Cvoid}, op_data::Any)

See libhdf5 documentation for H5Ewalk2.

H5F — File Interface

h5f_close

h5f_create

h5f_flush

h5f_get_access_plist

h5f_get_create_plist

h5f_get_intent

h5f_get_name

h5f_get_obj_count

h5f_get_obj_ids

h5f_get_vfd_handle

h5f_is_hdf5

h5f_open

h5f_start_swmr_write

HDF5.API.h5f_close — Function

h5f_close(file_id::hid_t)

See libhdf5 documentation for H5Fclose.

HDF5.API.h5f_create — Function

h5f_create(pathname::Ptr{UInt8}, flags::Cuint, fcpl_id::hid_t, fapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Fcreate.

HDF5.API.h5f_flush — Function

h5f_flush(object_id::hid_t, scope::Cint)

See libhdf5 documentation for H5Fflush.

HDF5.API.h5f_get_access_plist — Function

h5f_get_access_plist(file_id::hid_t) -> hid_t

See libhdf5 documentation for H5Fget_access_plist.

HDF5.API.h5f_get_create_plist — Function

h5f_get_create_plist(file_id::hid_t) -> hid_t

See libhdf5 documentation for H5Fget_create_plist.

HDF5.API.h5f_get_intent — Function

h5f_get_intent(file_id::hid_t, intent::Ptr{Cuint})

See libhdf5 documentation for H5Fget_intent.

HDF5.API.h5f_get_name — Function

h5f_get_name(obj_id::hid_t, buf::Ptr{UInt8}, buf_size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5Fget_name.

HDF5.API.h5f_get_obj_count — Function

h5f_get_obj_count(file_id::hid_t, types::Cuint) -> Cssize_t

See libhdf5 documentation for H5Fget_obj_count.

HDF5.API.h5f_get_obj_ids — Function

h5f_get_obj_ids(file_id::hid_t, types::Cuint, max_objs::Csize_t, obj_id_list::Ptr{hid_t}) -> Cssize_t

See libhdf5 documentation for H5Fget_obj_ids.

HDF5.API.h5f_get_vfd_handle — Function

h5f_get_vfd_handle(file_id::hid_t, fapl_id::hid_t, file_handle::Ref{Ptr{Cvoid}})

See libhdf5 documentation for H5Fget_vfd_handle.

HDF5.API.h5f_is_hdf5 — Function

h5f_is_hdf5(pathname::Cstring) -> Bool

See libhdf5 documentation for H5Fis_hdf5.

HDF5.API.h5f_open — Function

h5f_open(pathname::Cstring, flags::Cuint, fapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Fopen.

HDF5.API.h5f_start_swmr_write — Function

h5f_start_swmr_write(id::hid_t)

See libhdf5 documentation for H5Fstart_swmr_write.

H5G — Group Interface

h5g_close

h5g_create

h5g_get_create_plist

h5g_get_info

h5g_get_num_objs

h5g_get_objname_by_idx

h5g_open

HDF5.API.h5g_close — Function

h5g_close(group_id::hid_t)

See libhdf5 documentation for H5Gclose.

HDF5.API.h5g_create — Function

h5g_create(loc_id::hid_t, pathname::Ptr{UInt8}, lcpl_id::hid_t, gcpl_id::hid_t, gapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Gcreate2.

HDF5.API.h5g_get_create_plist — Function

h5g_get_create_plist(group_id::hid_t) -> hid_t

See libhdf5 documentation for H5Gget_create_plist.

HDF5.API.h5g_get_info — Function

h5g_get_info(group_id::hid_t, buf::Ptr{H5G_info_t})

See libhdf5 documentation for H5Gget_info.

HDF5.API.h5g_get_num_objs — Function

h5g_get_num_objs(loc_id::hid_t, num_obj::Ptr{hsize_t}) -> hid_t

See libhdf5 documentation for H5Gget_num_objs.

HDF5.API.h5g_get_objname_by_idx — Function

h5g_get_objname_by_idx(loc_id::hid_t, idx::hsize_t, pathname::Ptr{UInt8}, size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5Gget_objname_by_idx.

HDF5.API.h5g_open — Function

h5g_open(loc_id::hid_t, pathname::Ptr{UInt8}, gapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Gopen2.

H5I — Identifier Interface

h5i_dec_ref

h5i_get_file_id

h5i_get_name

h5i_get_ref

h5i_get_type

h5i_inc_ref

h5i_is_valid

HDF5.API.h5i_dec_ref — Function

h5i_dec_ref(obj_id::hid_t) -> Int

See libhdf5 documentation for H5Idec_ref.

HDF5.API.h5i_get_file_id — Function

h5i_get_file_id(obj_id::hid_t) -> hid_t

See libhdf5 documentation for H5Iget_file_id.

HDF5.API.h5i_get_name — Function

h5i_get_name(obj_id::hid_t, buf::Ptr{UInt8}, buf_size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5Iget_name.

HDF5.API.h5i_get_ref — Function

h5i_get_ref(obj_id::hid_t) -> Int

See libhdf5 documentation for H5Iget_ref.

HDF5.API.h5i_get_type — Function

h5i_get_type(obj_id::hid_t) -> Int

See libhdf5 documentation for H5Iget_type.

HDF5.API.h5i_inc_ref — Function

h5i_inc_ref(obj_id::hid_t) -> Int

See libhdf5 documentation for H5Iinc_ref.

HDF5.API.h5i_is_valid — Function

h5i_is_valid(obj_id::hid_t) -> Bool

See libhdf5 documentation for H5Iis_valid.

H5L — Link Interface

h5l_create_external

h5l_create_hard

h5l_create_soft

h5l_delete

h5l_exists

h5l_get_info

h5l_get_name_by_idx

h5l_iterate

h5l_move

HDF5.API.h5l_create_external — Function

h5l_create_external(target_file_name::Ptr{UInt8}, target_obj_name::Ptr{UInt8}, link_loc_id::hid_t, link_name::Ptr{UInt8}, lcpl_id::hid_t, lapl_id::hid_t)

See libhdf5 documentation for H5Lcreate_external.

HDF5.API.h5l_create_hard — Function

h5l_create_hard(obj_loc_id::hid_t, obj_name::Ptr{UInt8}, link_loc_id::hid_t, link_name::Ptr{UInt8}, lcpl_id::hid_t, lapl_id::hid_t)

See libhdf5 documentation for H5Lcreate_hard.

HDF5.API.h5l_create_soft — Function

h5l_create_soft(target_path::Ptr{UInt8}, link_loc_id::hid_t, link_name::Ptr{UInt8}, lcpl_id::hid_t, lapl_id::hid_t)

See libhdf5 documentation for H5Lcreate_soft.

HDF5.API.h5l_delete — Function

h5l_delete(obj_id::hid_t, pathname::Ptr{UInt8}, lapl_id::hid_t)

See libhdf5 documentation for H5Ldelete.

HDF5.API.h5l_exists — Function

h5l_exists(loc_id::hid_t, pathname::Ptr{UInt8}, lapl_id::hid_t) -> Bool

See libhdf5 documentation for H5Lexists.

HDF5.API.h5l_get_info — Function

h5l_get_info(link_loc_id::hid_t, link_name::Ptr{UInt8}, link_buf::Ptr{H5L_info_t}, lapl_id::hid_t)

See libhdf5 documentation for H5Lget_info.

HDF5.API.h5l_get_name_by_idx — Function

h5l_get_name_by_idx(loc_id::hid_t, group_name::Ptr{UInt8}, index_field::Cint, order::Cint, n::hsize_t, name::Ptr{UInt8}, size::Csize_t, lapl_id::hid_t) -> Cssize_t

See libhdf5 documentation for H5Lget_name_by_idx.

HDF5.API.h5l_iterate — Function

h5l_iterate(group_id::hid_t, idx_type::Cint, order::Cint, idx::Ptr{hsize_t}, op::Ptr{Cvoid}, op_data::Any)

See libhdf5 documentation for H5Literate1.

h5l_iterate(f, group_id, idx_type, order, idx = 0) -> hsize_t

Executes h5l_iterate with the user-provided callback function f, returning the index where iteration ends.

The callback function must correspond to the signature

    f(group::HDF5.API.hid_t, name::Ptr{Cchar}, info::Ptr{HDF5.API.H5L_info_t}) -> HDF5.API.herr_t

where a negative return value halts iteration abnormally, a positive value halts iteration successfully, and zero continues iteration.

Examples

julia> HDF5.API.h5l_iterate(hfile, HDF5.API.H5_INDEX_NAME, HDF5.API.H5_ITER_INC) do group, name, info
           println(unsafe_string(name))
           return HDF5.API.herr_t(0)
       end

HDF5.API.h5l_move — Function

h5l_move(src_obj_id::hid_t, src_name::Ptr{UInt8}, dest_obj_id::hid_t, dest_name::Ptr{UInt8}, lcpl_id::hid_t, lapl_id::hid_t)

See libhdf5 documentation for H5Lmove.

H5O — Object Interface

h5o_close

h5o_copy

h5o_get_info

h5o_open

h5o_open_by_addr

h5o_open_by_idx

HDF5.API.h5o_close — Function

h5o_close(object_id::hid_t)

See libhdf5 documentation for H5Oclose.

HDF5.API.h5o_copy — Function

h5o_copy(src_loc_id::hid_t, src_name::Ptr{UInt8}, dst_loc_id::hid_t, dst_name::Ptr{UInt8}, ocpypl_id::hid_t, lcpl_id::hid_t)

See libhdf5 documentation for H5Ocopy.

HDF5.API.h5o_get_info — Function

h5o_get_info(object_id::hid_t, buf::Ptr{H5O_info_t})

See libhdf5 documentation for H5Oget_info1.

HDF5.API.h5o_open — Function

h5o_open(loc_id::hid_t, pathname::Ptr{UInt8}, lapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Oopen.

HDF5.API.h5o_open_by_addr — Function

h5o_open_by_addr(loc_id::hid_t, addr::haddr_t) -> hid_t

See libhdf5 documentation for H5Oopen_by_addr.

HDF5.API.h5o_open_by_idx — Function

h5o_open_by_idx(loc_id::hid_t, group_name::Ptr{UInt8}, index_type::Cint, order::Cint, n::hsize_t, lapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Oopen_by_idx.

H5PL — Plugin Interface

h5pl_append

h5pl_get

h5pl_get_loading_state

h5pl_insert

h5pl_prepend

h5pl_remove

h5pl_replace

h5pl_set_loading_state

h5pl_size

HDF5.API.h5pl_append — Function

h5pl_append(search_path::Ptr{Cchar})

See libhdf5 documentation for H5PLappend.

HDF5.API.h5pl_get — Function

h5pl_get(index::Cuint, path_buf::Ptr{Cchar}, buf_size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5PLget.

HDF5.API.h5pl_get_loading_state — Function

h5pl_get_loading_state(plugin_control_mask::Ptr{Cuint})

See libhdf5 documentation for H5PLget_loading_state.

HDF5.API.h5pl_insert — Function

h5pl_insert(search_path::Ptr{Cchar}, index::Cuint)

See libhdf5 documentation for H5PLinsert.

HDF5.API.h5pl_prepend — Function

h5pl_prepend(search_path::Ptr{Cchar})

See libhdf5 documentation for H5PLprepend.

HDF5.API.h5pl_remove — Function

h5pl_remove(index::Cuint)

See libhdf5 documentation for H5PLremove.

HDF5.API.h5pl_replace — Function

h5pl_replace(search_path::Ptr{Cchar}, index::Cuint)

See libhdf5 documentation for H5PLreplace.

HDF5.API.h5pl_set_loading_state — Function

h5pl_set_loading_state(plugin_control_mask::Cuint)

See libhdf5 documentation for H5PLset_loading_state.

HDF5.API.h5pl_size — Function

h5pl_size(num_paths::Ptr{Cuint})

See libhdf5 documentation for H5PLsize.

H5P — Property Interface

h5p_close

h5p_create

h5p_get_alignment

h5p_get_alloc_time

h5p_get_attr_creation_order

h5p_get_char_encoding

h5p_get_chunk

h5p_get_class_name

h5p_get_core_write_tracking

h5p_get_create_intermediate_group

h5p_get_driver

h5p_get_driver_info

h5p_get_dxpl_mpio

h5p_get_efile_prefix

h5p_get_external

h5p_get_external_count

h5p_get_fapl_core

h5p_get_fapl_family

h5p_get_fapl_hdfs

h5p_get_fapl_mpio32

h5p_get_fapl_mpio64

h5p_get_fapl_multi

h5p_get_fapl_splitter

h5p_get_fclose_degree

h5p_get_fill_time

h5p_get_fill_value

h5p_get_filter

h5p_get_filter_by_id

h5p_get_layout

h5p_get_libver_bounds

h5p_get_link_creation_order

h5p_get_local_heap_size_hint

h5p_get_nfilters

h5p_get_obj_track_times

h5p_get_userblock

h5p_modify_filter

h5p_remove_filter

h5p_set_alignment

h5p_set_alloc_time

h5p_set_attr_creation_order

h5p_set_char_encoding

h5p_set_chunk

h5p_set_chunk_cache

h5p_set_core_write_tracking

h5p_set_create_intermediate_group

h5p_set_deflate

h5p_set_dxpl_mpio

h5p_set_efile_prefix

h5p_set_external

h5p_set_fapl_core

h5p_set_fapl_family

h5p_set_fapl_hdfs

h5p_set_fapl_log

h5p_set_fapl_mpio32

h5p_set_fapl_mpio64

h5p_set_fapl_multi

h5p_set_fapl_sec2

h5p_set_fapl_split

h5p_set_fapl_splitter

h5p_set_fapl_windows

h5p_set_fclose_degree

h5p_set_fill_time

h5p_set_fill_value

h5p_set_filter

h5p_set_fletcher32

h5p_set_layout

h5p_set_libver_bounds

h5p_set_link_creation_order

h5p_set_local_heap_size_hint

h5p_set_nbit

h5p_set_obj_track_times

h5p_set_scaleoffset

h5p_set_shuffle

h5p_set_szip

h5p_set_userblock

h5p_set_virtual

HDF5.API.h5p_close — Function

h5p_close(id::hid_t)

See libhdf5 documentation for H5Pclose.

HDF5.API.h5p_create — Function

h5p_create(cls_id::hid_t) -> hid_t

See libhdf5 documentation for H5Pcreate.

HDF5.API.h5p_get_alignment — Function

h5p_get_alignment(fapl_id::hid_t, threshold::Ref{hsize_t}, alignment::Ref{hsize_t})

See libhdf5 documentation for H5Pget_alignment.

HDF5.API.h5p_get_alloc_time — Function

h5p_get_alloc_time(plist_id::hid_t, alloc_time::Ptr{Cint})

See libhdf5 documentation for H5Pget_alloc_time.

HDF5.API.h5p_get_attr_creation_order — Function

h5p_get_attr_creation_order(plist_id::hid_t, crt_order_flags::Ptr{Cuint})

See libhdf5 documentation for H5Pget_attr_creation_order.

HDF5.API.h5p_get_char_encoding — Function

h5p_get_char_encoding(plist_id::hid_t, encoding::Ref{Cint})

See libhdf5 documentation for H5Pget_char_encoding.

HDF5.API.h5p_get_chunk — Function

h5p_get_chunk(plist_id::hid_t, n_dims::Cint, dims::Ptr{hsize_t}) -> Int

See libhdf5 documentation for H5Pget_chunk.

HDF5.API.h5p_get_class_name — Function

h5p_get_class_name(pcid::hid_t) -> String

See libhdf5 documentation for H5P_GET_CLASS_NAME.

HDF5.API.h5p_get_core_write_tracking — Function

h5p_get_core_write_tracking(fapl_id::hid_t, is_enabled::Ptr{hbool_t}, page_size::Ptr{Csize_t})

See libhdf5 documentation for H5Pget_core_write_tracking.

HDF5.API.h5p_get_create_intermediate_group — Function

h5p_get_create_intermediate_group(lcpl_id::hid_t, crt_intermed_group::Ref{Cuint})

See libhdf5 documentation for H5Pget_create_intermediate_group.

HDF5.API.h5p_get_driver — Function

h5p_get_driver(plist_id::hid_t) -> hid_t

See libhdf5 documentation for H5Pget_driver.

HDF5.API.h5p_get_driver_info — Function

h5p_get_driver_info(plist_id::hid_t) -> Ptr{Cvoid}

See libhdf5 documentation for H5Pget_driver_info.

HDF5.API.h5p_get_dxpl_mpio — Function

h5p_get_dxpl_mpio(dxpl_id::hid_t, xfer_mode::Ptr{Cint})

See libhdf5 documentation for H5Pget_dxpl_mpio.

HDF5.API.h5p_get_efile_prefix — Function

h5p_get_efile_prefix(dapl_id::hid_t, prefix::Ptr{UInt8}, size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5Pget_efile_prefix.

HDF5.API.h5p_get_external — Function

h5p_get_external(plist::hid_t, idx::Cuint, name_size::Csize_t, name::Ptr{Cuchar}, offset::Ptr{off_t}, size::Ptr{hsize_t})

See libhdf5 documentation for H5Pget_external.

HDF5.API.h5p_get_external_count — Function

h5p_get_external_count(plist::hid_t) -> Int

See libhdf5 documentation for H5Pget_external_count.

HDF5.API.h5p_get_fapl_core — Function

h5p_get_fapl_core(fapl_id::hid_t, increment::Ptr{Csize_t}, backing_store::Ptr{hbool_t})

See libhdf5 documentation for H5Pget_fapl_core.

HDF5.API.h5p_get_fapl_family — Function

h5p_get_fapl_family(fapl_id::hid_t, memb_size::Ptr{hsize_t}, memb_fapl_id::Ptr{hid_t})

See libhdf5 documentation for H5Pget_fapl_family.

HDF5.API.h5p_get_fapl_hdfs — Function

h5p_get_fapl_hdfs(fapl_id::hid_t, fa_out::Ptr{H5FD_hdfs_fapl_t})

See libhdf5 documentation for H5Pget_fapl_hdfs.

HDF5.API.h5p_get_fapl_mpio32 — Function

h5p_get_fapl_mpio32(fapl_id::hid_t, comm::Ptr{Hmpih32}, info::Ptr{Hmpih32})

See libhdf5 documentation for H5Pget_fapl_mpio.

HDF5.API.h5p_get_fapl_mpio64 — Function

h5p_get_fapl_mpio64(fapl_id::hid_t, comm::Ptr{Hmpih64}, info::Ptr{Hmpih64})

See libhdf5 documentation for H5Pget_fapl_mpio.

HDF5.API.h5p_get_fapl_multi — Function

h5p_get_fapl_multi(fapl_id::hid_t, memb_map::Ptr{H5FD_mem_t}, memb_fapl::Ptr{hid_t}, memb_name::Ptr{Ptr{Cchar}}, memb_addr::Ptr{haddr_t}, relax::Ptr{hbool_t})

See libhdf5 documentation for H5Pget_fapl_multi.

HDF5.API.h5p_get_fapl_splitter — Function

h5p_get_fapl_splitter(fapl_id::hid_t, config_ptr::Ptr{H5FD_splitter_vfd_config_t})

See libhdf5 documentation for H5Pget_fapl_splitter.

HDF5.API.h5p_get_fclose_degree — Function

h5p_get_fclose_degree(fapl_id::hid_t, fc_degree::Ref{Cint})

See libhdf5 documentation for H5Pget_fclose_degree.

HDF5.API.h5p_get_fill_time — Function

h5p_get_fill_time(plist_id::hid_t, fill_time::Ptr{H5D_fill_time_t})

See libhdf5 documentation for H5Pget_fill_time.

HDF5.API.h5p_get_fill_value — Function

h5p_get_fill_value(plist_id::hid_t, type_id::hid_t, value::Ptr{Cvoid})

See libhdf5 documentation for H5Pget_fill_value.

HDF5.API.h5p_get_filter — Function

h5p_get_filter(plist_id::hid_t, idx::Cuint, flags::Ptr{Cuint}, cd_nemlts::Ref{Csize_t}, cd_values::Ptr{Cuint}, namelen::Csize_t, name::Ptr{Cchar}, filter_config::Ptr{Cuint}) -> H5Z_filter_t

See libhdf5 documentation for H5Pget_filter2.

HDF5.API.h5p_get_filter_by_id — Function

h5p_get_filter_by_id(plist_id::hid_t, filter_id::H5Z_filter_t, flags::Ref{Cuint}, cd_nelmts::Ref{Csize_t}, cd_values::Ptr{Cuint}, namelen::Csize_t, name::Ptr{UInt8}, filter_config::Ptr{Cuint})

See libhdf5 documentation for H5Pget_filter_by_id2.

HDF5.API.h5p_get_layout — Function

h5p_get_layout(plist_id::hid_t) -> Int

See libhdf5 documentation for H5Pget_layout.

HDF5.API.h5p_get_libver_bounds — Function

h5p_get_libver_bounds(fapl_id::hid_t, low::Ref{Cint}, high::Ref{Cint})

See libhdf5 documentation for H5Pget_libver_bounds.

HDF5.API.h5p_get_link_creation_order — Function

h5p_get_link_creation_order(plist_id::hid_t, crt_order_flags::Ptr{Cuint})

See libhdf5 documentation for H5Pget_link_creation_order.

HDF5.API.h5p_get_local_heap_size_hint — Function

h5p_get_local_heap_size_hint(plist_id::hid_t, size_hint::Ref{Csize_t})

See libhdf5 documentation for H5Pget_local_heap_size_hint.

HDF5.API.h5p_get_nfilters — Function

h5p_get_nfilters(plist_id::hid_t) -> Int

See libhdf5 documentation for H5Pget_nfilters.

HDF5.API.h5p_get_obj_track_times — Function

h5p_get_obj_track_times(plist_id::hid_t, track_times::Ref{UInt8})

See libhdf5 documentation for H5Pget_obj_track_times.

HDF5.API.h5p_get_userblock — Function

h5p_get_userblock(plist_id::hid_t, len::Ptr{hsize_t})

See libhdf5 documentation for H5Pget_userblock.

HDF5.API.h5p_modify_filter — Function

h5p_modify_filter(plist_id::hid_t, filter_id::H5Z_filter_t, flags::Cuint, cd_nelmts::Csize_t, cd_values::Ptr{Cuint})

See libhdf5 documentation for H5Pmodify_filter.

HDF5.API.h5p_remove_filter — Function

h5p_remove_filter(plist_id::hid_t, filter_id::H5Z_filter_t)

See libhdf5 documentation for H5Premove_filter.

HDF5.API.h5p_set_alignment — Function

h5p_set_alignment(plist_id::hid_t, threshold::hsize_t, alignment::hsize_t)

See libhdf5 documentation for H5Pset_alignment.

HDF5.API.h5p_set_alloc_time — Function

h5p_set_alloc_time(plist_id::hid_t, alloc_time::Cint)

See libhdf5 documentation for H5Pset_alloc_time.

HDF5.API.h5p_set_attr_creation_order — Function

h5p_set_attr_creation_order(plist_id::hid_t, crt_order_flags::Cuint)

See libhdf5 documentation for H5Pset_attr_creation_order.

HDF5.API.h5p_set_char_encoding — Function

h5p_set_char_encoding(plist_id::hid_t, encoding::Cint)

See libhdf5 documentation for H5Pset_char_encoding.

HDF5.API.h5p_set_chunk — Function

h5p_set_chunk(plist_id::hid_t, ndims::Cint, dims::Ptr{hsize_t})

See libhdf5 documentation for H5Pset_chunk.

HDF5.API.h5p_set_chunk_cache — Function

h5p_set_chunk_cache(dapl_id::hid_t, rdcc_nslots::Csize_t, rdcc_nbytes::Csize_t, rdcc_w0::Cdouble)

See libhdf5 documentation for H5Pset_chunk_cache.

HDF5.API.h5p_set_core_write_tracking — Function

h5p_set_core_write_tracking(fapl_id::hid_t, is_enabled::hbool_t, page_size::Csize_t)

See libhdf5 documentation for H5Pset_core_write_tracking.

HDF5.API.h5p_set_create_intermediate_group — Function

h5p_set_create_intermediate_group(plist_id::hid_t, setting::Cuint)

See libhdf5 documentation for H5Pset_create_intermediate_group.

HDF5.API.h5p_set_deflate — Function

h5p_set_deflate(plist_id::hid_t, setting::Cuint)

See libhdf5 documentation for H5Pset_deflate.

HDF5.API.h5p_set_dxpl_mpio — Function

h5p_set_dxpl_mpio(dxpl_id::hid_t, xfer_mode::Cint)

See libhdf5 documentation for H5Pset_dxpl_mpio.

HDF5.API.h5p_set_efile_prefix — Function

h5p_set_efile_prefix(plist_id::hid_t, prefix::Ptr{UInt8})

See libhdf5 documentation for H5Pset_efile_prefix.

HDF5.API.h5p_set_external — Function

h5p_set_external(plist_id::hid_t, name::Ptr{UInt8}, offset::off_t, size::Csize_t)

See libhdf5 documentation for H5Pset_external.

HDF5.API.h5p_set_fapl_core — Function

h5p_set_fapl_core(fapl_id::hid_t, increment::Csize_t, backing_store::hbool_t)

See libhdf5 documentation for H5Pset_fapl_core.

HDF5.API.h5p_set_fapl_family — Function

h5p_set_fapl_family(fapl_id::hid_t, memb_size::hsize_t, memb_fapl_id::hid_t)

See libhdf5 documentation for H5Pset_fapl_family.

HDF5.API.h5p_set_fapl_hdfs — Function

h5p_set_fapl_hdfs(fapl_id::hid_t, fa::Ptr{H5FD_hdfs_fapl_t})

See libhdf5 documentation for H5Pset_fapl_hdfs.

HDF5.API.h5p_set_fapl_log — Function

h5p_set_fapl_log(fapl_id::hid_t, logfile::Ptr{Cchar}, flags::Culonglong, buf_size::Csize_t)

See libhdf5 documentation for H5Pset_fapl_log.

HDF5.API.h5p_set_fapl_mpio32 — Function

h5p_set_fapl_mpio32(fapl_id::hid_t, comm::Hmpih32, info::Hmpih32)

See libhdf5 documentation for H5Pset_fapl_mpio.

HDF5.API.h5p_set_fapl_mpio64 — Function

h5p_set_fapl_mpio64(fapl_id::hid_t, comm::Hmpih64, info::Hmpih64)

See libhdf5 documentation for H5Pset_fapl_mpio.

HDF5.API.h5p_set_fapl_multi — Function

h5p_set_fapl_multi(fapl_id::hid_t, memb_map::Ptr{H5FD_mem_t}, memb_fapl::Ptr{hid_t}, memb_name::Ptr{Ptr{Cchar}}, memb_addr::Ptr{haddr_t}, relax::hbool_t)

See libhdf5 documentation for H5Pset_fapl_multi.

HDF5.API.h5p_set_fapl_sec2 — Function

h5p_set_fapl_sec2(fapl_id::hid_t)

See libhdf5 documentation for H5Pset_fapl_sec2.

HDF5.API.h5p_set_fapl_split — Function

h5p_set_fapl_split(fapl::hid_t, meta_ext::Ptr{Cchar}, meta_plist_id::hid_t, raw_ext::Ptr{Cchar}, raw_plist_id::hid_t)

See libhdf5 documentation for H5Pset_fapl_split.

HDF5.API.h5p_set_fapl_splitter — Function

h5p_set_fapl_splitter(fapl_id::hid_t, config_ptr::Ptr{H5FD_splitter_vfd_config_t})

See libhdf5 documentation for H5Pset_fapl_splitter.

HDF5.API.h5p_set_fapl_windows — Function

h5p_set_fapl_windows(fapl_id::hid_t)

See libhdf5 documentation for H5Pset_fapl_windows.

HDF5.API.h5p_set_fclose_degree — Function

h5p_set_fclose_degree(plist_id::hid_t, fc_degree::Cint)

See libhdf5 documentation for H5Pset_fclose_degree.

HDF5.API.h5p_set_fill_time — Function

h5p_set_fill_time(plist_id::hid_t, fill_time::H5D_fill_time_t)

See libhdf5 documentation for H5Pset_fill_time.

HDF5.API.h5p_set_fill_value — Function

h5p_set_fill_value(plist_id::hid_t, type_id::hid_t, value::Ptr{Cvoid})

See libhdf5 documentation for H5Pset_fill_value.

HDF5.API.h5p_set_filter — Function

h5p_set_filter(plist_id::hid_t, filter_id::H5Z_filter_t, flags::Cuint, cd_nelmts::Csize_t, cd_values::Ptr{Cuint})

See libhdf5 documentation for H5Pset_filter.

HDF5.API.h5p_set_fletcher32 — Function

h5p_set_fletcher32(plist_id::hid_t)

See libhdf5 documentation for H5Pset_fletcher32.

HDF5.API.h5p_set_layout — Function

h5p_set_layout(plist_id::hid_t, setting::Cint)

See libhdf5 documentation for H5Pset_layout.

HDF5.API.h5p_set_libver_bounds — Function

h5p_set_libver_bounds(fapl_id::hid_t, low::Cint, high::Cint)

See libhdf5 documentation for H5Pset_libver_bounds.

HDF5.API.h5p_set_link_creation_order — Function

h5p_set_link_creation_order(plist_id::hid_t, crt_order_flags::Cuint)

See libhdf5 documentation for H5Pset_link_creation_order.

HDF5.API.h5p_set_local_heap_size_hint — Function

h5p_set_local_heap_size_hint(plist_id::hid_t, size_hint::Csize_t)

See libhdf5 documentation for H5Pset_local_heap_size_hint.

HDF5.API.h5p_set_nbit — Function

h5p_set_nbit(plist_id::hid_t)

See libhdf5 documentation for H5Pset_nbit.

HDF5.API.h5p_set_obj_track_times — Function

h5p_set_obj_track_times(plist_id::hid_t, track_times::UInt8)

See libhdf5 documentation for H5Pset_obj_track_times.

HDF5.API.h5p_set_scaleoffset — Function

h5p_set_scaleoffset(plist_id::hid_t, scale_type::Cint, scale_factor::Cint)

See libhdf5 documentation for H5Pset_scaleoffset.

HDF5.API.h5p_set_shuffle — Function

h5p_set_shuffle(plist_id::hid_t)

See libhdf5 documentation for H5Pset_shuffle.

HDF5.API.h5p_set_szip — Function

h5p_set_szip(plist_id::hid_t, options_mask::Cuint, pixels_per_block::Cuint)

See libhdf5 documentation for H5Pset_szip.

HDF5.API.h5p_set_userblock — Function

h5p_set_userblock(plist_id::hid_t, len::hsize_t)

See libhdf5 documentation for H5Pset_userblock.

HDF5.API.h5p_set_virtual — Function

h5p_set_virtual(dcpl_id::hid_t, vspace_id::hid_t, src_file_name::Ptr{UInt8}, src_dset_name::Ptr{UInt8}, src_space_id::hid_t)

See libhdf5 documentation for H5Pset_virtual.

H5R — Reference Interface

h5r_create

h5r_dereference

h5r_get_obj_type

h5r_get_region

HDF5.API.h5r_create — Function

h5r_create(ref::Ptr{Cvoid}, loc_id::hid_t, pathname::Ptr{UInt8}, ref_type::Cint, space_id::hid_t)

See libhdf5 documentation for H5Rcreate.

HDF5.API.h5r_dereference — Function

h5r_dereference(obj_id::hid_t, oapl_id::hid_t, ref_type::Cint, ref::Ptr{Cvoid}) -> hid_t

See libhdf5 documentation for H5Rdereference2.

HDF5.API.h5r_get_obj_type — Function

h5r_get_obj_type(loc_id::hid_t, ref_type::Cint, ref::Ptr{Cvoid}, obj_type::Ptr{Cint})

See libhdf5 documentation for H5Rget_obj_type2.

HDF5.API.h5r_get_region — Function

h5r_get_region(loc_id::hid_t, ref_type::Cint, ref::Ptr{Cvoid}) -> hid_t

See libhdf5 documentation for H5Rget_region.

H5S — Dataspace Interface

h5s_close

h5s_combine_select

h5s_copy

h5s_create

h5s_create_simple

h5s_extent_equal

h5s_get_regular_hyperslab

h5s_get_select_hyper_nblocks

h5s_get_select_npoints

h5s_get_select_type

h5s_get_simple_extent_dims

h5s_get_simple_extent_ndims

h5s_get_simple_extent_type

h5s_is_regular_hyperslab

h5s_is_simple

h5s_select_hyperslab

h5s_set_extent_simple

HDF5.API.h5s_close — Function

h5s_close(space_id::hid_t)

See libhdf5 documentation for H5Sclose.

HDF5.API.h5s_combine_select — Function

h5s_combine_select(space1_id::hid_t, op::Cint, space2_id::hid_t) -> hid_t

See libhdf5 documentation for H5Scombine_select.

HDF5.API.h5s_copy — Function

h5s_copy(space_id::hid_t) -> hid_t

See libhdf5 documentation for H5Scopy.

HDF5.API.h5s_create — Function

h5s_create(class::Cint) -> hid_t

See libhdf5 documentation for H5Screate.

HDF5.API.h5s_create_simple — Function

h5s_create_simple(rank::Cint, current_dims::Ptr{hsize_t}, maximum_dims::Ptr{hsize_t}) -> hid_t

See libhdf5 documentation for H5Screate_simple.

HDF5.API.h5s_extent_equal — Function

h5s_extent_equal(space1_id::hid_t, space2_id::hid_t) -> Bool

See libhdf5 documentation for H5Sextent_equal.

HDF5.API.h5s_get_regular_hyperslab — Function

h5s_get_regular_hyperslab(space_id::hid_t, start::Ptr{hsize_t}, stride::Ptr{hsize_t}, count::Ptr{hsize_t}, block::Ptr{hsize_t})

See libhdf5 documentation for H5Sget_regular_hyperslab.

HDF5.API.h5s_get_select_hyper_nblocks — Function

h5s_get_select_hyper_nblocks(space_id::hid_t) -> hssize_t

See libhdf5 documentation for H5Sget_select_hyper_nblocks.

HDF5.API.h5s_get_select_npoints — Function

h5s_get_select_npoints(space_id::hid_t) -> hsize_t

See libhdf5 documentation for H5Sget_select_npoints.

HDF5.API.h5s_get_select_type — Function

h5s_get_select_type(space_id::hid_t) -> Int

See libhdf5 documentation for H5Sget_select_type.

HDF5.API.h5s_get_simple_extent_dims — Function

h5s_get_simple_extent_dims(space_id::hid_t, dims::Ptr{hsize_t}, maxdims::Ptr{hsize_t}) -> Int

See libhdf5 documentation for H5Sget_simple_extent_dims.

HDF5.API.h5s_get_simple_extent_ndims — Function

h5s_get_simple_extent_ndims(space_id::hid_t) -> Int

See libhdf5 documentation for H5Sget_simple_extent_ndims.

HDF5.API.h5s_get_simple_extent_type — Function

h5s_get_simple_extent_type(space_id::hid_t) -> Int

See libhdf5 documentation for H5Sget_simple_extent_type.

HDF5.API.h5s_is_regular_hyperslab — Function

h5s_is_regular_hyperslab(space_id::hid_t) -> Bool

See libhdf5 documentation for H5Sis_regular_hyperslab.

HDF5.API.h5s_is_simple — Function

h5s_is_simple(space_id::hid_t) -> Bool

See libhdf5 documentation for H5Sis_simple.

HDF5.API.h5s_select_hyperslab — Function

h5s_select_hyperslab(dspace_id::hid_t, seloper::Cint, start::Ptr{hsize_t}, stride::Ptr{hsize_t}, count::Ptr{hsize_t}, block::Ptr{hsize_t})

See libhdf5 documentation for H5Sselect_hyperslab.

HDF5.API.h5s_set_extent_simple — Function

h5s_set_extent_simple(dspace_id::hid_t, rank::Cint, current_size::Ptr{hsize_t}, maximum_size::Ptr{hsize_t})

See libhdf5 documentation for H5Sset_extent_simple.

H5T — Datatype Interface

h5t_array_create

h5t_close

h5t_commit

h5t_committed

h5t_copy

h5t_create

h5t_enum_insert

h5t_equal

h5t_get_array_dims

h5t_get_array_ndims

h5t_get_class

h5t_get_cset

h5t_get_ebias

h5t_get_fields

h5t_get_member_class

h5t_get_member_index

h5t_get_member_name

h5t_get_member_offset

h5t_get_member_type

h5t_get_native_type

h5t_get_nmembers

h5t_get_offset

h5t_get_order

h5t_get_precision

h5t_get_sign

h5t_get_size

h5t_get_strpad

h5t_get_super

h5t_get_tag

h5t_insert

h5t_is_variable_str

h5t_lock

h5t_open

h5t_set_cset

h5t_set_ebias

h5t_set_fields

h5t_set_offset

h5t_set_order

h5t_set_precision

h5t_set_size

h5t_set_strpad

h5t_set_tag

h5t_vlen_create

HDF5.API.h5t_array_create — Function

h5t_array_create(basetype_id::hid_t, ndims::Cuint, sz::Ptr{hsize_t}) -> hid_t

See libhdf5 documentation for H5Tarray_create2.

HDF5.API.h5t_close — Function

h5t_close(dtype_id::hid_t)

See libhdf5 documentation for H5Tclose.

HDF5.API.h5t_commit — Function

h5t_commit(loc_id::hid_t, name::Ptr{UInt8}, dtype_id::hid_t, lcpl_id::hid_t, tcpl_id::hid_t, tapl_id::hid_t)

See libhdf5 documentation for H5Tcommit2.

HDF5.API.h5t_committed — Function

h5t_committed(dtype_id::hid_t) -> Bool

See libhdf5 documentation for H5Tcommitted.

HDF5.API.h5t_copy — Function

h5t_copy(dtype_id::hid_t) -> hid_t

See libhdf5 documentation for H5Tcopy.

HDF5.API.h5t_create — Function

h5t_create(class_id::Cint, sz::Csize_t) -> hid_t

See libhdf5 documentation for H5Tcreate.

HDF5.API.h5t_enum_insert — Function

h5t_enum_insert(dtype_id::hid_t, name::Cstring, value::Ptr{Cvoid})

See libhdf5 documentation for H5Tenum_insert.

HDF5.API.h5t_equal — Function

h5t_equal(dtype_id1::hid_t, dtype_id2::hid_t) -> Bool

See libhdf5 documentation for H5Tequal.

HDF5.API.h5t_get_array_dims — Function

h5t_get_array_dims(dtype_id::hid_t, dims::Ptr{hsize_t}) -> Int

See libhdf5 documentation for H5Tget_array_dims2.

HDF5.API.h5t_get_array_ndims — Function

h5t_get_array_ndims(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_array_ndims.

HDF5.API.h5t_get_class — Function

h5t_get_class(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_class.

HDF5.API.h5t_get_cset — Function

h5t_get_cset(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_cset.

HDF5.API.h5t_get_ebias — Function

h5t_get_ebias(dtype_id::hid_t) -> Csize_t

See libhdf5 documentation for H5Tget_ebias.

HDF5.API.h5t_get_fields — Function

h5t_get_fields(dtype_id::hid_t, spos::Ref{Csize_t}, epos::Ref{Csize_t}, esize::Ref{Csize_t}, mpos::Ref{Csize_t}, msize::Ref{Csize_t})

See libhdf5 documentation for H5Tget_fields.

HDF5.API.h5t_get_member_class — Function

h5t_get_member_class(dtype_id::hid_t, index::Cuint) -> Int

See libhdf5 documentation for H5Tget_member_class.

HDF5.API.h5t_get_member_index — Function

h5t_get_member_index(dtype_id::hid_t, membername::Ptr{UInt8}) -> Int

See libhdf5 documentation for H5Tget_member_index.

HDF5.API.h5t_get_member_name — Function

h5t_get_member_name(type_id::hid_t, index::Cuint) -> String

See libhdf5 documentation for H5Oopen.

HDF5.API.h5t_get_member_offset — Function

h5t_get_member_offset(dtype_id::hid_t, index::Cuint) -> Csize_t

See libhdf5 documentation for H5Tget_member_offset.

HDF5.API.h5t_get_member_type — Function

h5t_get_member_type(dtype_id::hid_t, index::Cuint) -> hid_t

See libhdf5 documentation for H5Tget_member_type.

HDF5.API.h5t_get_native_type — Function

h5t_get_native_type(dtype_id::hid_t, direction::Cint) -> hid_t

See libhdf5 documentation for H5Tget_native_type.

HDF5.API.h5t_get_nmembers — Function

h5t_get_nmembers(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_nmembers.

HDF5.API.h5t_get_offset — Function

h5t_get_offset(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_offset.

HDF5.API.h5t_get_order — Function

h5t_get_order(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_order.

HDF5.API.h5t_get_precision — Function

h5t_get_precision(dtype_id::hid_t) -> Csize_t

See libhdf5 documentation for H5Tget_precision.

HDF5.API.h5t_get_sign — Function

h5t_get_sign(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_sign.

HDF5.API.h5t_get_size — Function

h5t_get_size(dtype_id::hid_t) -> Csize_t

See libhdf5 documentation for H5Tget_size.

HDF5.API.h5t_get_strpad — Function

h5t_get_strpad(dtype_id::hid_t) -> Int

See libhdf5 documentation for H5Tget_strpad.

HDF5.API.h5t_get_super — Function

h5t_get_super(dtype_id::hid_t) -> hid_t

See libhdf5 documentation for H5Tget_super.

HDF5.API.h5t_get_tag — Function

h5t_get_tag(type_id::hid_t) -> String

See libhdf5 documentation for H5Oopen.

HDF5.API.h5t_insert — Function

h5t_insert(dtype_id::hid_t, fieldname::Ptr{UInt8}, offset::Csize_t, field_id::hid_t)

See libhdf5 documentation for H5Tinsert.

HDF5.API.h5t_is_variable_str — Function

h5t_is_variable_str(type_id::hid_t) -> Bool

See libhdf5 documentation for H5Tis_variable_str.

HDF5.API.h5t_lock — Function

h5t_lock(type_id::hid_t)

See libhdf5 documentation for H5Tlock.

HDF5.API.h5t_open — Function

h5t_open(loc_id::hid_t, name::Ptr{UInt8}, tapl_id::hid_t) -> hid_t

See libhdf5 documentation for H5Topen2.

HDF5.API.h5t_set_cset — Function

h5t_set_cset(dtype_id::hid_t, cset::Cint)

See libhdf5 documentation for H5Tset_cset.

HDF5.API.h5t_set_ebias — Function

h5t_set_ebias(dtype_id::hid_t, ebias::Csize_t)

See libhdf5 documentation for H5Tset_ebias.

HDF5.API.h5t_set_fields — Function

h5t_set_fields(dtype_id::hid_t, spos::Csize_t, epos::Csize_t, esize::Csize_t, mpos::Csize_t, msize::Csize_t)

See libhdf5 documentation for H5Tset_fields.

HDF5.API.h5t_set_offset — Function

h5t_set_offset(dtype_id::hid_t, offset::Csize_t)

See libhdf5 documentation for H5Tset_offset.

HDF5.API.h5t_set_order — Function

h5t_set_order(dtype_id::hid_t, order::Cint)

See libhdf5 documentation for H5Tset_order.

HDF5.API.h5t_set_precision — Function

h5t_set_precision(dtype_id::hid_t, sz::Csize_t)

See libhdf5 documentation for H5Tset_precision.

HDF5.API.h5t_set_size — Function

h5t_set_size(dtype_id::hid_t, sz::Csize_t)

See libhdf5 documentation for H5Tset_size.

HDF5.API.h5t_set_strpad — Function

h5t_set_strpad(dtype_id::hid_t, sz::Cint)

See libhdf5 documentation for H5Tset_strpad.

HDF5.API.h5t_set_tag — Function

h5t_set_tag(dtype_id::hid_t, tag::Cstring)

See libhdf5 documentation for H5Tset_tag.

HDF5.API.h5t_vlen_create — Function

h5t_vlen_create(base_type_id::hid_t) -> hid_t

See libhdf5 documentation for H5Tvlen_create.

H5Z — Filter Interface

h5z_filter_avail

h5z_get_filter_info

h5z_register

h5z_unregister

HDF5.API.h5z_filter_avail — Function

h5z_filter_avail(id::H5Z_filter_t) -> Bool

See libhdf5 documentation for H5Zfilter_avail.

HDF5.API.h5z_get_filter_info — Function

h5z_get_filter_info(filter::H5Z_filter_t, filter_config_flags::Ptr{Cuint})

See libhdf5 documentation for H5Zget_filter_info.

HDF5.API.h5z_register — Function

h5z_register(filter_class::Ref{H5Z_class_t})

See libhdf5 documentation for H5Zregister.

HDF5.API.h5z_unregister — Function

h5z_unregister(id::H5Z_filter_t)

See libhdf5 documentation for H5Zunregister.

H5DO — Optimized Functions Interface

h5do_append

h5do_write_chunk

HDF5.API.h5do_append — Function

h5do_append(dset_id::hid_t, dxpl_id::hid_t, index::Cuint, num_elem::hsize_t, memtype::hid_t, buffer::Ptr{Cvoid})

See libhdf5 documentation for H5DOappend.

HDF5.API.h5do_write_chunk — Function

h5do_write_chunk(dset_id::hid_t, dxpl_id::hid_t, filter_mask::UInt32, offset::Ptr{hsize_t}, bufsize::Csize_t, buf::Ptr{Cvoid})

See libhdf5 documentation for H5DOwrite_chunk.

H5DS — Dimension Scale Interface

h5ds_attach_scale

h5ds_detach_scale

h5ds_get_label

h5ds_get_num_scales

h5ds_get_scale_name

h5ds_is_attached

h5ds_is_scale

h5ds_set_label

h5ds_set_scale

HDF5.API.h5ds_attach_scale — Function

h5ds_attach_scale(did::hid_t, dsid::hid_t, idx::Cuint)

See libhdf5 documentation for H5DSattach_scale.

HDF5.API.h5ds_detach_scale — Function

h5ds_detach_scale(did::hid_t, dsid::hid_t, idx::Cuint)

See libhdf5 documentation for H5DSdetach_scale.

HDF5.API.h5ds_get_label — Function

h5ds_get_label(did::hid_t, idx::Cuint, label::Ptr{UInt8}, size::hsize_t)

See libhdf5 documentation for H5DSget_label.

HDF5.API.h5ds_get_num_scales — Function

h5ds_get_num_scales(did::hid_t, idx::Cuint) -> Int

See libhdf5 documentation for H5DSget_num_scales.

HDF5.API.h5ds_get_scale_name — Function

h5ds_get_scale_name(did::hid_t, name::Ptr{UInt8}, size::Csize_t) -> Cssize_t

See libhdf5 documentation for H5DSget_scale_name.

HDF5.API.h5ds_is_attached — Function

h5ds_is_attached(did::hid_t, dsid::hid_t, idx::Cuint) -> Bool

See libhdf5 documentation for H5DSis_attached.

HDF5.API.h5ds_is_scale — Function

h5ds_is_scale(did::hid_t) -> Bool

See libhdf5 documentation for H5DSis_scale.

HDF5.API.h5ds_set_label — Function

h5ds_set_label(did::hid_t, idx::Cuint, label::Ref{UInt8})

See libhdf5 documentation for H5DSset_label.

HDF5.API.h5ds_set_scale — Function

h5ds_set_scale(dsid::hid_t, dimname::Ptr{UInt8})

See libhdf5 documentation for H5DSset_scale.

H5LT — Lite Interface

h5lt_dtype_to_text

HDF5.API.h5lt_dtype_to_text — Function

h5lt_dtype_to_text(datatype::hid_t, str::Ptr{UInt8}, lang_type::Cint, len::Ref{Csize_t})

See libhdf5 documentation for H5LTdtype_to_text.

H5TB — Table Interface

h5tb_append_records

h5tb_get_field_info

h5tb_get_table_info

h5tb_make_table

h5tb_read_records

h5tb_read_table

h5tb_write_records

HDF5.API.h5tb_append_records — Function

h5tb_append_records(loc_id::hid_t, dset_name::Ptr{UInt8}, nrecords::hsize_t, type_size::Csize_t, field_offset::Ptr{Csize_t}, field_sizes::Ptr{Csize_t}, data::Ptr{Cvoid})

See libhdf5 documentation for H5TBappend_records.

HDF5.API.h5tb_get_field_info — Function

h5tb_get_field_info(loc_id::hid_t, table_name::Ptr{UInt8}, field_names::Ptr{Ptr{UInt8}}, field_sizes::Ptr{Csize_t}, field_offsets::Ptr{Csize_t}, type_size::Ptr{Csize_t})

See libhdf5 documentation for H5TBget_field_info.

HDF5.API.h5tb_get_table_info — Function

h5tb_get_table_info(loc_id::hid_t, table_name::Ptr{UInt8}, nfields::Ptr{hsize_t}, nrecords::Ptr{hsize_t})

See libhdf5 documentation for H5TBget_table_info.

HDF5.API.h5tb_make_table — Function

h5tb_make_table(table_title::Ptr{UInt8}, loc_id::hid_t, dset_name::Ptr{UInt8}, nfields::hsize_t, nrecords::hsize_t, type_size::Csize_t, field_names::Ptr{Ptr{UInt8}}, field_offset::Ptr{Csize_t}, field_types::Ptr{hid_t}, chunk_size::hsize_t, fill_data::Ptr{Cvoid}, compress::Cint, data::Ptr{Cvoid})

See libhdf5 documentation for H5TBmake_table.

HDF5.API.h5tb_read_records — Function

h5tb_read_records(loc_id::hid_t, table_name::Ptr{UInt8}, start::hsize_t, nrecords::hsize_t, type_size::Csize_t, field_offsets::Ptr{Csize_t}, dst_sizes::Ptr{Csize_t}, data::Ptr{Cvoid})

See libhdf5 documentation for H5TBread_records.

HDF5.API.h5tb_read_table — Function

h5tb_read_table(loc_id::hid_t, table_name::Ptr{UInt8}, dst_size::Csize_t, dst_offset::Ptr{Csize_t}, dst_sizes::Ptr{Csize_t}, dst_buf::Ptr{Cvoid})

See libhdf5 documentation for H5TBread_table.

HDF5.API.h5tb_write_records — Function

h5tb_write_records(loc_id::hid_t, table_name::Ptr{UInt8}, start::hsize_t, nrecords::hsize_t, type_size::Csize_t, field_offsets::Ptr{UInt8}, field_sizes::Ptr{UInt8}, data::Ptr{Cvoid})

See libhdf5 documentation for H5TBwrite_records.