Dataspaces

HDF5.Dataspace — Type

HDF5.Dataspace

A dataspace defines the size and the shape of a dataset or an attribute.

A dataspace is typically constructed by calling dataspace.

The following functions have methods defined for Dataspace objects

==

ndims

size

length

isempty

isnull

HDF5.dataspace — Function

dataspace(obj::Union{Attribute, Dataset, Dataspace})

The Dataspace of obj.

dataspace(data)

The default Dataspace used for representing a Julia object data:

strings or numbers: a scalar Dataspace

arrays: a simple Dataspace

nothing or an EmptyArray: a null dataspace

dataspace(dims::Tuple; maxdims::Tuple=dims)

Construct a simple Dataspace for the given dimensions dims. The maximum dimensions maxdims specifies the maximum possible size: -1 can be used to indicate unlimited dimensions.

HDF5.isnull — Function

isnull(dspace::Union{HDF5.Dataspace, HDF5.Dataset, HDF5.Attribute})

Determines whether the given object has no size (consistent with the API.H5S_NULL dataspace).

Examples

julia> HDF5.isnull(dataspace(HDF5.EmptyArray{Float64}()))
true

julia> HDF5.isnull(dataspace((0,)))
false