Filters

HDF5 supports filters for compression and validation: these are applied sequentially to each chunk of a dataset when writing data, and in reverse order when reading data.

These can be set by passing a filter or vector of filters as a filters property to DatasetCreateProperties or via the filters keyword argument of create_dataset.

Example

HDF5.Filters — Module

HDF5.Filters

This module contains the interface for using filters in HDF5.jl.

Example Usage

using HDF5
using HDF5.Filters

# Create a new file
fn = tempname()

# Create test data
data = rand(1000, 1000)

# Open temp file for writing
f = h5open(fn, "w") 

# Create datasets
dsdeflate = create_dataset(f, "deflate", datatype(data), dataspace(data),
                           chunk=(100, 100), deflate=3)

dsshufdef = create_dataset(f, "shufdef", datatype(data), dataspace(data),
                           chunk=(100, 100), shuffle=true, deflate=3)

dsfiltdef = create_dataset(f, "filtdef", datatype(data), dataspace(data),
                           chunk=(100, 100), filters=Filters.Deflate(3))

dsfiltshufdef = create_dataset(f, "filtshufdef", datatype(data), dataspace(data),
                               chunk=(100, 100), filters=[Filters.Shuffle(), Filters.Deflate(3)])

# Write data
write(dsdeflate, data)
write(dsshufdef, data)
write(dsfiltdef, data)
write(dsfiltshufdef, data)

close(f)

Additonal Examples

See test/filter.jl for further examples.

Built-in Filters

HDF5.Filters.Deflate — Type

Deflate(level=5)

Deflate/ZLIB lossless compression filter. level is an integer between 0 and 9, inclusive, denoting the compression level, with 0 being no compression, 9 being the highest compression (but slowest speed).

External links

H5P_SET_DEFLATE

Deflate on Wikipedia

HDF5.Filters.Shuffle — Type

Shuffle()

The shuffle filter de-interlaces a block of data by reordering the bytes. All the bytes from one consistent byte position of each data element are placed together in one block; all bytes from a second consistent byte position of each data element are placed together a second block; etc. For example, given three data elements of a 4-byte datatype stored as 012301230123, shuffling will re-order data as 000111222333. This can be a valuable step in an effective compression algorithm because the bytes in each byte position are often closely related to each other and putting them together can increase the compression ratio.

As implied above, the primary value of the shuffle filter lies in its coordinated use with a compression filter; it does not provide data compression when used alone. When the shuffle filter is applied to a dataset immediately prior to the use of a compression filter, the compression ratio achieved is often superior to that achieved by the use of a compression filter without the shuffle filter.

External links

H5P_SET_SHUFFLE

HDF5.Filters.Fletcher32 — Type

Fletcher32()

The Fletcher32 checksum filter. This doesn't perform compression, but instead checks the validity of the stored data.

This should be applied after any lossy filters have been applied.

External links

H5P_SET_FLETCHER32

Fletcher's checksum on Wikipedia

HDF5.Filters.Szip — Type

Szip(coding=:nn, pixels_per_block=8)

Szip compression lossless filter. Options:

coding: the coding method: either :ec (entropy coding) or :nn (nearest neighbors, default)

pixels_per_block: The number of pixels or data elements in each data block (typically 8, 10, 16, or 32)

External links

H5P_SET_SZIP

Szip Compression in HDF Products

HDF5.Filters.NBit — Type

NBit()

The N-Bit filter.

External links

H5P_SET_NBIT

HDF5.Filters.ScaleOffset — Type

ScaleOffset(scale_type::Integer, scale_offset::Integer)

The scale-offset filter.

External links

H5P_SET_SCALEOFFSET

HDF5.Filters.UnknownFilter — Type

UnknownFilter(filter_id::API.H5Z_filter_t, flags::Cuint, data::Vector{Cuint}, name::String, config::Cuint)

An unknown filter.

External Filter Packages

Several external Julia packages implement HDF5 filter plugins in Julia. As they are independent of HDF5.jl, they must be installed in order to use their plugins.

The H5Zblosc.jl, H5Zbzip2.jl, H5Zlz4.jl, and H5Zzstd.jl packages are maintained as independent subdirectory packages within the HDF5.jl repository.

H5Zblosc.jl

H5Zblosc.BloscFilter — Type

BloscFilter(;level=5, shuffle=true, compressor="blosclz")

The Blosc compression filter, using Blosc.jl. Options:

level: compression level

shuffle: whether to shuffle data before compressing (this option should be used instead of the Shuffle filter)

compressor: the compression algorithm. Call Blosc.compressors() for the available compressors.

External links

What Is Blosc?

H5Zbzip2.jl

Missing docstring.

Missing docstring for Bzip2Filter. Check Documenter's build log for details.

H5Zlz4.jl

Missing docstring.

Missing docstring for Lz4Filter. Check Documenter's build log for details.

H5Zzstd.jl

Missing docstring.

Missing docstring for ZstdFilter. Check Documenter's build log for details.

Other External Filters

Additional filters can be dynamically loaded by the HDF5 library. See External Links below for more information.

Filter Interface

The filter interface is used to describe filters and obtain information on them.

HDF5.Filters.Filter — Type

Filter

Abstract type to describe HDF5 Filters. See the Extended Help for information on implementing a new filter.

Extended Help

Filter interface

The Filter interface is implemented upon the Filter subtype.

See API.h5z_register for details.

Required Methods to Implement

filterid - registered filter ID

filter_func - implement the actual filter

Optional Methods to Implement

filtername - defaults to "Unnamed Filter"

encoder_present - defaults to true

decoder_present - defaults to true

can_apply_func - defaults to nothing

set_local_func - defaults to nothing

Advanced Methods to Implement

can_apply_cfunc - Defaults to wrapping @cfunction around the result of can_apply_func

set_local_cfunc - Defaults to wrapping @cfunction around the result of set_local_func

filter_cfunc - Defaults to wrapping @cfunction around the result of filter_func

register_filter - Defaults to using the above functions to register the filter

Implement the Advanced Methods to avoid @cfunction from generating a runtime closure which may not work on all systems.

HDF5.Filters.FilterPipeline — Type

FilterPipeline(plist::DatasetCreateProperties)

The filter pipeline associated with plist. Acts like a AbstractVector{Filter}, supporting the following operations:

length(pipeline): the number of filters.

pipeline[i] to return the ith filter.

pipeline[FilterType] to return a filter of type FilterType

push!(pipline, filter) to add an extra filter to the pipeline.

append!(pipeline, filters) to add multiple filters to the pipeline.

delete!(pipeline, FilterType) to remove a filter of type FilterType from the pipeline.

empty!(pipeline) to remove all filters from the pipeline.

HDF5.Filters.FILTERS — Constant

FILTERS

Maps filter id to filter type.

HDF5.Filters.filterid — Function

filterid(F) where {F <: Filter}

The internal filter id of a filter of type F.

HDF5.Filters.encoder_present — Function

encoder_present(::Type{F}) where {F<:Filter}

Can the filter have an encode or compress the data? Defaults to true. Returns a Bool. See API.h5z_register.

HDF5.Filters.decoder_present — Function

decoder_present(::Type{F}) where {F<:Filter}

Can the filter decode or decompress the data? Defaults to true. Returns a Bool. See API.h5z_register

HDF5.Filters.filtername — Function

filtername(::Type{F}) where {F<:Filter}

What is the name of a filter? Defaults to "Unnamed Filter" Returns a String describing the filter. See API.h5z_register

HDF5.Filters.can_apply_func — Function

can_apply_func(::Type{F}) where {F<:Filter}

Return a function indicating whether the filter can be applied or nothing if no function exists. The function signature is func(dcpl_id::API.hid_t, type_id::API.hid_t, space_id::API.hid_t). See API.h5z_register

HDF5.Filters.can_apply_cfunc — Function

can_apply_cfunc(::Type{F}) where {F<:Filter}

Return a C function pointer for the can apply function. By default, this will return the result of using @cfunction on the function specified by can_apply_func(F) or C_NULL if nothing.

Overriding this will allow @cfunction to return a Ptr{Nothing} rather than a CFunction` closure which may not work on all systems.

HDF5.Filters.set_local_func — Function

set_local_func(::Type{F}) where {F<:Filter}

Return a function that sets dataset specific parameters or nothing if no function exists. The function signature is func(dcpl_id::API.hid_t, type_id::API.hid_t, space_id::API.hid_t). See API.h5z_register.

HDF5.Filters.set_local_cfunc — Function

set_local_cfunc(::Type{F}) where {F<:Filter}

Return a C function pointer for the set local function. By default, this will return the result of using @cfunction on the function specified by set_local_func(F) or C_NULL if nothing.

Overriding this will allow @cfunction to return a Ptr{Nothing} rather than a CFunction` closure which may not work on all systems.

HDF5.Filters.filter_func — Function

filter_func(::Type{F}) where {F<:Filter}

Returns a function that performs the actual filtering.

See API.h5z_register

HDF5.Filters.filter_cfunc — Function

filter_cfunc(::Type{F}) where {F<:Filter}

Return a C function pointer for the filter function. By default, this will return the result of using @cfunction on the function specified by filter_func(F) or will throw an error if nothing.

Overriding this will allow @cfunction to return a Ptr{Nothing} rather than a CFunction` closure which may not work on all systems.

HDF5.Filters.register_filter — Function

register_filter(::Type{F}) where F <: Filter

Register the filter with the HDF5 library via API.h5z_register. Also add F to the FILTERS dictionary.

External Links

A list of registered filter plugins can be found on the HDF Group website.

See the HDF5 Documentation of HDF5 Filter Plugins for details.

The source code for many external plugins have been collected in the HDFGroup hdf5_plugins repository.

Compiled binaries of dynamically downloaded plugins by downloaded from HDF5 Group.