Configuration

HDF5.has_parallel — Function

has_parallel()

Returns true if the HDF5 libraries were compiled with MPI parallel support via the Drivers.MPIO driver.

See Parallel HDF5 for more details.

HDF5.has_ros3 — Function

has_ros3()

Returns true if the HDF5 libraries were compiled with ros3 support

Display

HDF5.SHOW_TREE_ICONS — Constant

SHOW_TREE_ICONS = Ref{Bool}(true)

Configurable option to control whether emoji icons (true) or a plain-text annotation (false) is used to indicate the object type by show_tree.

HDF5.SHOW_TREE_MAX_CHILDREN — Constant

SHOW_TREE_MAX_CHILDREN = Ref{Int}(50)

Maximum number of children to show at each node.

HDF5.SHOW_TREE_MAX_DEPTH — Constant

SHOW_TREE_MAX_DEPTH = Ref{Int}(5)

Maximum recursive depth to descend during printing.

Internals

HDF5.get_context_property — Function

get_context_property(name::Symbol)

Internal API

Retrieve a property list from the task local context, defaulting to HDF5.CONTEXT if task_local_storage()[:hdf5_context] does not exist.

HDF5.CONTEXT — Constant

HDF5.CONTEXT

Internal API

Default HDF5Context.

HDF5.HDF5Context — Type

HDF5Context

Internal API

An HDF5Context is a collection of HDF5 property lists. It is meant to be used as a Task local mechanism to store state and change the default property lists for new objects.

Use the function get_context_property(name::Symbol) to access a property list within the local context.

The context in task_local_storage()[:hdf5_context] will be checked first. A common global HDF5Context is stored in the constant HDF5.CONTEXT and serves as the default context if the current task does not have a :hdf5_context.

Fields

attribute_access

attribute_create

dataset_access

dataset_create

dataset_tranfer

datatype_access

datatype_create

file_access

file_create

file_mount

group_access

group_create

link_access

link_create

object_copy

object_create

string_create