Dataspaces

Missing docstring.

Missing docstring for Dataspace. Check Documenter's build log for details.

HDF5.dataspace — Function

dataspace(obj::Union{Attribute, Dataset, Dataspace})

The Dataspace of obj.

dataspace(data)

The default Dataspace used for representing a Julia object data:

strings or numbers: a scalar Dataspace

arrays: a simple Dataspace

struct types: a scalar Dataspace

nothing or an EmptyArray: a null dataspace

dataspace(dims::Tuple; maxdims::Tuple=dims)

Construct a simple Dataspace for the given dimensions dims. The maximum dimensions maxdims specifies the maximum possible size: -1 can be used to indicate unlimited dimensions.

Missing docstring.

Missing docstring for isnull. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BlockRange. Check Documenter's build log for details.

Missing docstring.

Missing docstring for select_hyperslab!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_regular_hyperslab. Check Documenter's build log for details.