Groups

HDF5.create_group — Function

create_group(parent::Union{File,Group}, path::AbstractString; properties...)

Create a new Group at path under the parent object. Optional keyword arguments include any keywords that that belong to LinkCreateProperties or GroupCreateProperties.

HDF5.open_group — Function

open_group(parent::Union{File,Group}, path::AbstratString)

Open an existing Group at path under the parent object.