Datatypes

HDF5.Datatype — Type

HDF5.Datatype(id, toclose = true)

Wrapper for a HDF5 datatype id. If toclose is true, the finalizer will close the datatype.

HDF5.open_datatype — Function

open_datatype(parent::Union{File,Group}, path::AbstractString; properties...)

Open an existing Datatype at path under the parent object.

Optional keyword arguments include any keywords that that belong to DatatypeAccessProperties.