Files

HDF5.h5open — Function

h5open(filename::AbstractString, mode::AbstractString="r"; swmr=false, pv...)

Open or create an HDF5 file where mode is one of:

"r" read only

"r+" read and write

"cw" read and write, create file if not existing, do not truncate

"w" read and write, create a new file (destroys any existing contents)

Pass swmr=true to enable (Single Writer Multiple Reader) SWMR write access for "w" and "r+", or SWMR read access for "r".

Properties can be specified as keywords for FileAccessProperties and FileCreateProperties.

Also the keywords fapl and fcpl can be used to provide default instances of these property lists. Property lists passed in via keyword will be closed. This is useful to set properties not currently defined by HDF5.jl.

Note that h5open uses fclose_degree = :strong by default, but this can be overriden by the fapl keyword.

function h5open(f::Function, args...; pv...)

Apply the function f to the result of h5open(args...; kwargs...) and close the resulting HDF5.File upon completion. For example with a do block:

h5open("foo.h5","w") do h5
    h5["foo"]=[1,2,3]
end

h5open(filename, [mode="r"], comm::MPI.Comm, [info::MPI.Info]; pv...)

Open or create a parallel HDF5 file using the MPI-IO driver.

Equivalent to h5open(filename, mode; fapl_mpio=(comm, info), pv...). Throws an informative error if the loaded HDF5 libraries do not include parallel support.

See the HDF5 docs for details on the comm and info arguments.

HDF5.ishdf5 — Function

ishdf5(name::AbstractString)

Returns true if the file specified by name is in the HDF5 format, and false otherwise.

Base.isopen — Function

isopen(obj::HDF5.File)

Returns true if obj has not been closed, false if it has been closed.

HDF5.start_swmr_write — Function

start_swmr_write(h5::HDF5.File)

Start Single Reader Multiple Writer (SWMR) writing mode.

External links

Single Writer Multiple Reader from the HDF5 manual.