Groups

HDF5.Group — Type

HDF5.Group

An object representing a HDF5 group. A group is analagous to a file system directory, in that, except for the root group, every object must be a member of at least one group.

See also

create_group

open_group

HDF5.create_group — Function

create_group(parent::Union{File,Group}, path::AbstractString; properties...)

Create a new Group at path under the parent object. Optional keyword arguments include any keywords that that belong to LinkCreateProperties or GroupCreateProperties.

HDF5.open_group — Function

open_group(parent::Union{File,Group}, path::AbstractString; properties...)

Open an existing Group at path under the parent object.

Optional keyword arguments include any keywords that that belong to GroupAccessProperties.

HDF5.create_external — Function

create_external(source::Union{HDF5.File, HDF5.Group}, source_relpath, target_filename, target_path;
                lcpl_id=HDF5.API.H5P_DEFAULT, lapl_id=HDF5.H5P.DEFAULT)

Create an external link such that source[source_relpath] points to target_path within the file with path target_filename.

See also

API.h5l_create_external